chore: run comments directly from action
This reduces complexity by removing the need to call a reusable workflow,
reducing the size of the payload that dispatches the workflow, and
reducing the number of version checks. It also allows us not to send the
source repo token in the JSON payload.

The ~downside is that teh action will use the pages URL that the user has
set, instead of the one given us directly by the Pages action.